Divert building method to container-based

+++ scripts/download.sh
@@ -5,7 +5,12 @@
 
 set -eo pipefail
 
-cd openwrt
+if [ -z "${OPENWRT_DIR}" -o -z "${OPENWRT_WORK_DIR}" ]; then
+  echo "'OPENWRT_DIR' or 'OPENWRT_WORK_DIR' is empty" >&2
+  exit 1
+fi
+
+cd "${OPENWRT_WORK_DIR}"
 make download -j8
 find dl -size -1024c -exec ls -l {} \;
 find dl -size -1024c -exec rm -f {} \;
